requirements.txt
numpy
opencv-python
scikit-image
matplotlib
seaborn
xgboost
scikit-learn
flask
Pillow 